No longer instantiate a CoefficientMatrix for every matrix.
This leads to a tremendous memory savings and a substantial speedup.
This required the moving of the getters/setters form
CoeffiicentMatrix to CoefficientPlanes.

Additionally moved several fields from 32-bit values to size_t to
avoid calculations that accidently demote to 32-bit.

Lastly, the optical flow distance average is now normallized to the
diagonal of the video so we can compare flow between different
resolution images.

+++ src/BaseNddiDisplay.h
@@ -31,17 +31,17 @@ namespace nddi {
         /**
          * Holds the Display Width
          */
-        int displayWidth_;
+        size_t displayWidth_;
 
         /**
          * Holds the Display Height
          */
-        int displayHeight_;
+        size_t displayHeight_;
 
         /**
          * Holds the number of coefficient planes
          */
-        int numPlanes_;
+        size_t numPlanes_;
 
         /**
          * The input vector
